提交service代码

+++ src/main/java/com/genture/simulatePlatform/model/RoadCondition.java
@@ -4,43 +4,43 @@ package com.genture.simulatePlatform.model;
  * Created by Administrator on 2018/1/5.
  */
 public class RoadCondition {
-	private long unimpededDistance;
-	private long ambleDistance;
-	private long congestionDistance;
+	private int unimpededDistance;
+	private int ambleDistance;
+	private int congestionDistance;
 	private boolean status = false;
 
 	public RoadCondition(){
 
 	}
 
-	public RoadCondition(long unimpededDistance, long ambleDistance, long congestionDistance, boolean status){
+	public RoadCondition(int unimpededDistance, int ambleDistance, int congestionDistance, boolean status){
 		this.ambleDistance = ambleDistance;
 		this.congestionDistance = congestionDistance;
 		this.unimpededDistance = unimpededDistance;
 		this.status = status;
 	}
 
-	public long getUnimpededDistance() {
+	public int getUnimpededDistance() {
 		return unimpededDistance;
 	}
 
-	public void setUnimpededDistance(long unimpededDistance) {
+	public void setUnimpededDistance(int unimpededDistance) {
 		this.unimpededDistance = unimpededDistance;
 	}
 
-	public long getAmbleDistance() {
+	public int getAmbleDistance() {
 		return ambleDistance;
 	}
 
-	public void setAmbleDistance(long ambleDistance) {
+	public void setAmbleDistance(int ambleDistance) {
 		this.ambleDistance = ambleDistance;
 	}
 
-	public long getCongestionDistance() {
+	public int getCongestionDistance() {
 		return congestionDistance;
 	}
 
-	public void setCongestionDistance(long congestionDistance) {
+	public void setCongestionDistance(int congestionDistance) {
 		this.congestionDistance = congestionDistance;
 	}
